requirements.txt
# requirements.txt

streamlit
gspread
pandas
openpyxl
python-docx
fpdf2